use correct UART mechanisms
The UART implementation was actually quite wrong.

The programmer can't directly control the SO, SI, SD, SC lines in UART mode. And
the send/receive data flags are read-only. Also for writing from the GBA the
code was checking the control line directly. First this isn't necessary as this
should already be implemented in hardware: the GBA won't send data as long as SC
is high. The code should monitor the send data flag instead.

The code example to reflect this. Now the code uses a 1 byte header to specify
the length of the transmission which is used by the renamed rcv_uart
function.

Additionally there's an added python script at repl/repl.py. This should become
an actual repl. Currently it's more of a WIP. But at least the code example
isn't broken and confusing.

The source/main.c file was also converted to unix format.