add start of uart protocol mode, which uses header for transmit
So far we've been looking for a newline to signal the end of a
transmission. What would be more robust and flexible is to use a header
format. So far the header just contains one byte of lenght, which constrains the
lenght of a message to 256 bytes. In the future we should make the length field
bigger and we should add crc checking and a message type field.