minor changes: le wire format, use tonc aliases for types
Using network wire format was convenient for one of the code conversions, but
the year is 2020, and in general it's just a pain in the butt in a little endian
world.

On the alias front, I don't want to ever type unsigned char ever again. Not
totally done as I want to convert all mentions of char to u8.