switch type and length field in Gbaser protocol, fix nasty bug
Tried to solve the gba sending garbage bytes, by 'smartly' taking advantage of
the garbage pattern, and switching around the type and length fields in the
protocol. As we have more control over the contents of the type field than the
length field. For now it looks like the junk was caused by a read earlier in the
program, which was trying to remove the junk. Oh, how ironic. This needs a bit
more field-testing.

In other news it turns out there was a bug where the gba was stalling on reading
messages bigger than 127 bytes. That's of course quite a conspicuous number. It
turns out that when deserializing I was left shifting a signed number. And so
the left-most one bit is sticky to preserve negative.

This all does mean that after some quick testing nothing seems to intemittantly
break, and I can declare in lieu of actual proper testing that the protocol is
pretty roboust so far.

Still need to implement timeouts and error correction on both sides. And this
needs to be documented.